\name{GuanRank-package}
\alias{GuanRank-package}
\alias{GuanRank}
\docType{package}
\title{
\packageTitle{GuanRank}
}
\description{
\packageDescription{GuanRank}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{GuanRank}
\packageIndices{GuanRank}
This package provides a fast C++ version of GuanRank
}
\author{
\packageAuthor{GuanRank}

Maintainer: \packageMaintainer{GuanRank}
}
\references{
http://guanlab.ccmb.med.umich.edu/yuanfang/Guan_rank.pdf
}



\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{

#guanrank("../../../data/sim_data.csv")


}
