\name{GuanRank-package}
\alias{GuanRank-package}
\alias{GuanRank}
\docType{package}
\title{
\packageTitle{GuanRank}
}
\description{
\packageDescription{GuanRank}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{GuanRank}
\packageIndices{GuanRank}
This package provide a R interface for GuanRank model
}
\author{
\packageAuthor{GuanRank}

Maintainer: \packageMaintainer{GuanRank}
}
\references{
http://guanlab.ccmb.med.umich.edu/yuanfang/Guan_rank.pdf
}

\keyword{ package }
\seealso{

}
\examples{

}
