\name{GuanRank-package}
\alias{GuanRank-package}
\alias{GuanRank}
\docType{package}
\title{
\packageTitle{GuanRank}
}
\description{
\packageDescription{GuanRank}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{GuanRank}
\packageIndices{GuanRank}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{GuanRank}

Maintainer: \packageMaintainer{GuanRank}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
